+++ src/main/components/PianoRollToolbar/PianoRollToolbar.tsx
@@ -1,5 +1,3 @@
-import { AppBar, Toolbar } from "@material-ui/core"
-import { makeStyles } from "@material-ui/styles"
 import { observer } from "mobx-react-lite"
 import React, { FC, useCallback } from "react"
 import styled from "styled-components"
@@ -7,6 +5,7 @@ import { useStores } from "../../hooks/useStores"
 import InstrumentBrowser from "../InstrumentBrowser/InstrumentBrowser"
 import { AutoScrollButton } from "../Toolbar/AutoScrollButton"
 import QuantizeSelector from "../Toolbar/QuantizeSelector/QuantizeSelector"
+import { Toolbar } from "../Toolbar/Toolbar"
 import { TrackListMenuButton } from "../TrackList/TrackListMenuButton"
 import { EventListButton } from "./EventListButton"
 import { InstrumentButton } from "./InstrumentButton"
@@ -14,16 +13,6 @@ import { PanSlider } from "./PanSlider"
 import { ToolSelector } from "./ToolSelector"
 import { TrackNameInput } from "./TrackNameInput"
 import { VolumeSlider } from "./VolumeSlider"
-
-const useStyles = makeStyles((theme) => ({
-  appBar: {
-    background: "var(--background-color)",
-    borderBottom: "1px solid var(--divider-color)",
-  },
-  title: {
-    marginRight: "1rem",
-  },
-}))
 
 const Spacer = styled.div`
   width: 1rem;
@@ -55,42 +44,38 @@ export const PianoRollToolbar: FC = observer(() => {
     pianoRollStore.isQuantizeEnabled = !pianoRollStore.isQuantizeEnabled
   }, [pianoRollStore])
 
-  const classes = useStyles({})
-
   if (selectedTrack === undefined) {
     return <></>
   }
 
   return (
-    <AppBar position="static" elevation={0} className={classes.appBar}>
-      <Toolbar variant="dense">
-        <TrackListMenuButton />
+    <Toolbar>
+      <TrackListMenuButton />
 
-        <TrackNameInput />
+      <TrackNameInput />
 
-        <EventListButton />
+      <EventListButton />
 
-        <Spacer />
+      <Spacer />
 
-        <InstrumentButton />
-        <InstrumentBrowser />
+      <InstrumentButton />
+      <InstrumentBrowser />
 
-        <VolumeSlider trackId={selectedTrackId} />
-        <PanSlider trackId={selectedTrackId} />
+      <VolumeSlider trackId={selectedTrackId} />
+      <PanSlider trackId={selectedTrackId} />
 
-        <FlexibleSpacer />
+      <FlexibleSpacer />
 
-        <ToolSelector />
+      <ToolSelector />
 
-        <QuantizeSelector
-          value={quantize}
-          enabled={isQuantizeEnabled}
-          onSelect={onSelectQuantize}
-          onClickSwitch={onClickQuantizeSwitch}
-        />
+      <QuantizeSelector
+        value={quantize}
+        enabled={isQuantizeEnabled}
+        onSelect={onSelectQuantize}
+        onClickSwitch={onClickQuantizeSwitch}
+      />
 
-        <AutoScrollButton onClick={onClickAutoScroll} selected={autoScroll} />
-      </Toolbar>
-    </AppBar>
+      <AutoScrollButton onClick={onClickAutoScroll} selected={autoScroll} />
+    </Toolbar>
   )
 })
